Refactor, add installer and shell script.
