+++ .gitignore
@@ -2,4 +2,4 @@ __pycache__
 venv/
 *.pyc
 var/
-
+solr-7.5.0/
