+++ backend/docs/solr.sha1
@@ -1,0 +1,1 @@
+029a9f7670b6cb8a31afd3ce000d4b163347c6aa *solr-7.5.0.tgz
